remove standalone, create module!!
we now have a a TL module to replace all the standalone scripts for processing shows/segments. instead of writing a whole new script for each show/segment, just write one short script, importing this module, define some attributes, and run it.

this is the first commit for this new paradigm.