add timestamp function
previously, 'timestamp' was a global variable. it would be assigned once at runtime. if, after the script was running for a while, it made a reference to the variable, it was using the time it was first assigned, which could have been a looooong time before.

now, there is a new function: get_timestamp which is called any time we need this variable. this means the timestamp is updated each time it is called.