support for breakaway, audio levels
you can now pass in a "breakaway" time (in seconds!). ffmpeg will stop outputting at this time. we need this function for PNS, so we can output the full 6 minute segment, as well as at the breakaway point at 3 minutes.

you can now pass in an int for 'ff_level', telling ffmpeg how you want to apply dynamic compression. because of the way this algorithm works, some files may need this set differently. -21 is usually a safe bet for TL files - which is why we're leaving the initial value there, but in some cases we need to change it.