more attribute type checks
fixed the check_length function. now, both attributes must be set for the function to run.

splitting the attrib. type checks into another function. you send it the attrib, type you want to check against, and, for now, a string to return back in the Exception. not finished with this, I am out of time for the day.