new strategy for environment variables
moving all the system environment variables (ev) into a separate file. this will allow us to keep track of them, remind us to set them on new PCs. it's better that they're all referenced in one central place.

updating readme file to match

BUT!! just discovered an issue: if an ev is referenced here but not actually declared in the PC, it will throw an error. and, since our new ev.py file is immediately imported to the main file when run, any script will never stop running. so, need to add a try block here in the future. i am about out of time for today. come back to this. should all of the ev references be inside the try block? if so, what should be inside the except block? a pass? printing a message? send notification? what if the ev that's missing is one of the variables in the notify function? shit. come back to this.

reformatting imports to match PEP 8