try block for ev import, improve file transfer check function
-import ev
putting our import of the ev file inside a try block. hopefully this will cue the user in case there's an ev that doesn't exist.

-check_file_transferred
now using if/else statements instead of a try block. it turns out using the old try block was not actually testing whether the file existed in the destination. this works better.

-new function
created new function to print to  screen to alert the user of error. old way was doing this in each case. now, just call the function. shorter and simpler. don't repeat yourself. I'm also thinking about reusability in the future - we'll now be able to call this function from other scripts. may be better to move this function outside the class. come back later