create requirements.txt, update readme
so I just realized we need a requirement.txt file to pip install everything on a fresh install of this.

ok, so the requirements.txt contains all python packages currently installed on the TL Prod PC. not all of these are needed for this talklib module! In fact, probably about a third of them are likely entirely superfluous.  But, listing only those packages needed by this module turned out to be quite tedious, and potentially unreliable. so, just as a safe bet, we're using this (for now).