fix self.show in syslog messages
we want all our syslog messages to be prepended with the name of the show/file they come from. previously, we had been using self.show in every call to the syslog function. now, self.show is built into this function, so we can just pass the actual message.

need to figure out a workaround when we want to call the syslog function from an outside script where self.show will not be defined.