split run function into multiple functions
finally!

this could cause a problem for some types of scripts for "local" shows where running the algorithm to find the file doesn't find anything if it doesn't exit. If that is the case, we're now returning an exception which will state that we didn't declare the local_file attribute. this will be confusing to debug, need a new solution asap.