changes to send_sms function
we are moving the send_sms function out of the class and into the separate utils module. this way we can more easily use it elsewhere without having to instantiate an object first.

also adding a send_call function to the utils module. this is for sending phone calls via Twilio.