use pytest fixtures instead of old manual method
instead of creating a function to instantiate a talklib.TLshow object and then call the function for each test, use a fixture and pass the fixture as an argument to the tests. this is how pytest recommends you do this sort of thing.

FYI I will work on changing this for ALL tests for talklib package, today or soon.