new method for environment variables
now putting the EVs into a class that can be imported into other files and the classes in those files can inherit from this class. the old way used global variables which worked fine but it wasn't too suitable for testing.

the main reason we're doing this is for testing. now that the TLShow class is inheriting from the EV class, we can easily change the EV attributes for testing. so, for example, we can now easily change the 'destinations' attribute to use local folders instead of actual environment variables pulled from whatever machine you're working on. this way, we can work on this software without having to declare ALL of these EVs. this makes things more portable.

I've been working on this for several days and tried using the unittest patch functionality but I could never get it working right. this seems to be a bit of a workaround but ultimately accomplishes the same goal.