create new mock.py file to use for mocking things
we need to be able to mock environment variables and some other things. creating this mock file to put everything in one place so we can import it from other files

want to take a moment to address the fixture. in the middle of the fixture decorator is the yield statement. everything before th yield is setting up what happens, and everything after is tearing it down. hope that makes sense. also, just to note, this fixture gets called each and every time a test function that uses it gets run. so this seems a little wasteful...may want to reconsider in the future.