add fixture to mock out EV class
EV class is imported into the show file. that's where we want to mock it. make it a fixture and pass to every test.

I think there is an easier way to do this without passing to every single test. can probably combine with the other fixture. just doing this for now to test