use ffmpeg-python instead of subprocess!!!
we're finally ditching the subprocess call to FFmpeg! ffmpeg-python is a long(ish)-supported python wrapper for ffmpeg.

for security and other reasons we shouldn't be using a long subprocess command with shell=True. now it's pure python and cleaner.

the other big positive to this approach is we can use if statements to add to the kwargs if needed. you can see the old way we were doing this which was very hackish.