refactor tests
it was inefficient (and rude) to call the permalink (PNS) for every test, but that's what it was doing per the pytest fixture.

with these changes, the file will only be downloaded once (generally, unless it's removed intentionally as part of a test).

still needs some work to remove the global 'input_file' variable which is still hackish. issue #18