set syslog level for send_notifications, indent issue
for the send_notifications() function, we're setting the default syslog level at error, since we mostly only call this function for real errors.

also, change the indentation for the check_length() function. instead of putting everything in an else block, just return from the if block early, then carry on if not